moved screen components to their own directory
